RES-777: avoid overlap with res-781
